corrected process info gathering
changed to PerfRawData so that processortime would be "total time run on
processor" instead of "percent time running on processor in current
moment".  actually looks correct now.  Removed elapsed time field as it
was not that useful and when you switch to PerfRawData it actually
becomes a timestamp, not elapsed time